pyproject.toml
[project]
name = "claim-process"
version = "0.1.0"
description = "Healthcare claims processing service with net fee calculation"
readme = "README.md"
authors = [
    { name = "Hrishikesh Kakkad", email = "username@users.noreply.github.com" }
]
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "sqlmodel>=0.0.14",
    "psycopg2-binary>=2.9.9",
    "python-multipart>=0.0.6",
    "pydantic>=2.5.2",
    "slowapi>=0.1.9",
    "redis>=5.0.1",
    "alembic>=1.13.0",
    "python-dateutil>=2.8.2",
    "httpx>=0.25.2",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "black>=23.12.1",
    "ruff>=0.1.9",
    "mypy>=1.8.0",
]

[project.scripts]
claim-process = "claim_process:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
